prove Tidal::Data->find's 3rd param is optional.
Previously we were always providing the optional 3rd param (epoch_time) to
T:D->find, because (a) when it's not provided, it defaults to time, viz:

  $epoch ||= time;

and (b) we don't know when this test file will be run; it's quite likely
that it'll be run long after the fixture data in the test is stale,
hence un->find'able.

"perldoc CORE" shows (a bit poorly) how to sieze control of Perl's
built-in functions.  I wrote a script a few months ago that shows
time() is subject to our control: see gist.github.com/2646897. See also
"Test::MockTimer".  Finally, another option altogether is to make the
date strings in the fixture data dynamic to the time the test is run.